services/bff/docker-requirements.txt
fastapi
httpx
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose[cryptography]>=3.3.0
passlib
requests